blackbox: enable blackbox  storaging fault log with pstore/ramoops
ohos inclusion
category:feature
issue: #I4WQEN
CVE: NA

-------------------------------

1.enable bbox storaging fault log with pstore/ramoops
2.provide the same ability for pstore/ramoops with pstore/blk

+++ include/linux/blackbox.h
@@ -50,7 +50,7 @@ struct error_info {
 
 struct fault_log_info {
 	char flag[8];  /* 8 is the length of the flag */
-	int len;  /* length of the kernel fault log */
+	size_t len;  /* length of the kernel fault log */
 	struct error_info info;
 };
